python/requirements.txt
langchain-community
langchain-huggingface
langchain-groq
sentence-transformers
faiss-cpu
transformers
torch
python-dotenv